pass actions into newState, and add initialState(state:)
In order to add an action parameter to newState it had to be split up
into initialState(state:) and newState(state:action:), because there is
no action to pass in with the initial state.

Sometimes it's just convenient to check the action in a subscriber. Not
everything belongs in state, and not all actions can be processed by
inspecting state. Transient state can now be passed to subscribers via
actions without polluting state with garbage that gets stale quickly.

This is less pure but very practical in real use. For example, if there
is some new data for your collection view you don't want to put the
changes required into state because they're one-offs. You also don't
want to necessarily diff entire collections every time state changes.